requirements.txt
Flask==2.1.3
Flask-PyMongo==2.3.0
flask-mongoengine
pandas==2.1.0
Werkzeug==2.2.2
coverage==7.1.0
pytest==7.2.1
gunicorn==20.1.0